pyproject.toml
[tool.poetry]
name = "library"
version = "0.1.0"
description = ""
authors = ["ferril <ferril.darkdiver@gmailc.om>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
psycopg2 = "^2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
